Clean up pytest suite in tests/logic_verification.
- Organized tests into subdirectories: `analysis`, `core`, `gui`, `instruments`.
- Consolidated FFT manager tests into `tests/logic_verification/core/test_fft_manager.py`.
- Renamed and moved `test_time_array_refactor.py` to `tests/logic_verification/analysis/test_time_array.py`.
- Moved GUI and Instrument tests to respective folders and updated `sys.path` imports.
- Deleted redundant tests: `test_utils_formatting.py` and `test_spectrum_buffer.py`.
