+++ tests/logic_verification/analysis/test_time_array.py
@@ -4,11 +4,11 @@ import sys
 import os
 
 # Add src to path
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../../')))
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../../../')))
 
 from src.core.analysis import _get_time_array, _get_reference_signals
 
-class TestTimeArrayRefactor(unittest.TestCase):
+class TestTimeArray(unittest.TestCase):
     def test_get_time_array_basic(self):
         """Test basic functionality of _get_time_array."""
         N = 100
